refactor: 연결 정보를 dto 패키지로 이동
