+++ src/main/java/nextstep/ladder/dto/Connections.java
@@ -1,4 +1,4 @@
-package nextstep.ladder.domain.line;
+package nextstep.ladder.dto;
 
 import java.util.List;
