+++ src/test/java/nextstep/ladder/domain/line/LineTest.java
@@ -1,5 +1,6 @@
 package nextstep.ladder.domain.line;
 
+import nextstep.ladder.dto.Connections;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
